feat: Change order confirmation page to be CMS driven (#2731)
Closes #2691
