+++ projects/storefrontlib/src/cms-components/checkout/checkout.module.ts
@@ -11,7 +11,6 @@ import { CheckoutProgressMobileBottomModule } from './components/checkout-progre
 import { CheckoutProgressMobileTopModule } from './components/checkout-progress/checkout-progress-mobile-top/checkout-progress-mobile-top.module';
 import { CheckoutProgressModule } from './components/checkout-progress/checkout-progress.module';
 import { DeliveryModeModule } from './components/delivery-mode/delivery-mode.module';
-import { OrderConfirmationModule } from './components/order-confirmation/order-confirmation.module';
 import { PaymentMethodModule } from './components/payment-method/payment-method.module';
 import { PlaceOrderModule } from './components/place-order/place-order.module';
 import { PromotionsModule } from './components/promotions/promotions.module';
@@ -29,7 +28,6 @@ import { ShippingAddressModule } from './components/shipping-address/shipping-ad
     CheckoutProgressMobileTopModule,
     CheckoutProgressMobileBottomModule,
     DeliveryModeModule,
-    OrderConfirmationModule,
     PaymentMethodModule,
     PlaceOrderModule,
     PromotionsModule,
