+++ projects/storefrontlib/src/cms-components/checkout/components/index.ts
@@ -10,8 +10,6 @@ export * from './checkout-progress/checkout-progress.component';
 export * from './checkout-progress/checkout-progress.module';
 export * from './delivery-mode/delivery-mode.component';
 export * from './delivery-mode/delivery-mode.module';
-export * from './order-confirmation/order-confirmation.component';
-export * from './order-confirmation/order-confirmation.module';
 export * from './payment-method/billing-address-form/billing-address-form.component';
 export * from './payment-method/billing-address-form/billing-address-form.module';
 export * from './payment-method/payment-form/payment-form.component';
