--- projects/storefrontlib/src/cms-components/checkout/components/order-confirmation/order-confirmation.component.html
@@ -1,88 +1,0 @@
-<div class="cx-page" *ngIf="(order$ | async) as order">
-  <header class="cx-page-header">
-    <h1 class="cx-page-title">
-      {{ 'checkoutOrderConfirmation.confirmationOfOrder' | cxTranslate }}
-      {{ order.code }}
-    </h1>
-  </header>
-
-  <div class="cx-order-confirmation">
-    <div class="cx-order-confirmation-message">
-      <h2>{{ 'checkoutOrderConfirmation.thankYou' | cxTranslate }}</h2>
-      <p>
-        {{
-          'checkoutOrderConfirmation.invoiceHasBeenSentByEmail' | cxTranslate
-        }}
-      </p>
-      <!-- <a href="#" class="btn-link">Print Page</a> -->
-    </div>
-
-    <cx-add-to-home-screen-banner></cx-add-to-home-screen-banner>
-
-    <div class="cx-order-review-summary">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-12 col-md-4 col-lg-3">
-            <div class="summary-card">
-              <cx-card
-                *ngIf="order.deliveryAddress"
-                [content]="getAddressCardContent(order.deliveryAddress) | async"
-              ></cx-card>
-            </div>
-          </div>
-
-          <div class="col-sm-12 col-md-4 col-lg-3">
-            <div class="summary-card">
-              <cx-card
-                *ngIf="order.paymentInfo"
-                [content]="
-                  getBillingAddressCardContent(
-                    order.paymentInfo.billingAddress
-                  ) | async
-                "
-              ></cx-card>
-            </div>
-          </div>
-
-          <div class="col-sm-12 col-md-4 col-lg-3">
-            <div class="summary-card">
-              <cx-card
-                *ngIf="order.deliveryMode"
-                [content]="
-                  getDeliveryModeCardContent(order.deliveryMode) | async
-                "
-              ></cx-card>
-            </div>
-          </div>
-
-          <div class="col-sm-12 col-md-4 col-lg-3">
-            <div class="summary-card">
-              <cx-card
-                *ngIf="order.paymentInfo"
-                [content]="getPaymentInfoCardContent(order.paymentInfo) | async"
-              ></cx-card>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div class="cx-order-items container">
-      <h4 class="cx-order-items-header">
-        {{ 'checkoutOrderConfirmation.orderItems' | cxTranslate }}
-      </h4>
-      <cx-cart-item-list
-        [items]="order.entries"
-        [isReadOnly]="true"
-      ></cx-cart-item-list>
-    </div>
-
-    <div class="cx-order-summary container">
-      <div class="row justify-content-end">
-        <div class="col-sm-12 col-md-6 col-lg-5 col-xl-4">
-          <cx-order-summary [cart]="order"></cx-order-summary>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
