+++ projects/storefrontlib/src/cms-components/checkout/guards/index.ts
@@ -1,4 +1,3 @@
-export * from './order-confirmation-page.guard';
 export * from './checkout.guard';
 export * from './delivery-mode-set.guard';
 export * from './shipping-address-set.guard';
