+++ projects/storefrontlib/src/cms-components/order-confirmation/components/index.ts
@@ -1,0 +1,4 @@
+export * from './order-confirmation-items/order-confirmation-items.component';
+export * from './order-confirmation-overview/order-confirmation-overview.component';
+export * from './order-confirmation-thank-you-message/order-confirmation-thank-you-message.component';
+export * from './order-confirmation-totals/order-confirmation-totals.component';
