+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-items/order-confirmation-items.component.spec.ts
@@ -1,0 +1,65 @@
+import { Component, Input } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { CheckoutService, I18nTestingModule, Order } from '@spartacus/core';
+import { Observable, of } from 'rxjs';
+import { Item } from '../../../cart/cart-shared/cart-item/cart-item.component';
+import { OrderConfirmationItemsComponent } from './order-confirmation-items.component';
+
+import createSpy = jasmine.createSpy;
+
+@Component({ selector: 'cx-cart-item-list', template: '' })
+class MockReviewSubmitComponent {
+  @Input()
+  items: Item[];
+  @Input()
+  isReadOnly: boolean;
+}
+
+class MockCheckoutService {
+  clearCheckoutData = createSpy();
+
+  getOrderDetails(): Observable<Order> {
+    return of({
+      entries: [
+        {
+          entryNumber: 1,
+          quantity: 1,
+        },
+      ],
+    });
+  }
+}
+
+describe('OrderConfirmationItemsComponent', () => {
+  let component: OrderConfirmationItemsComponent;
+  let fixture: ComponentFixture<OrderConfirmationItemsComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [I18nTestingModule],
+      declarations: [
+        OrderConfirmationItemsComponent,
+        MockReviewSubmitComponent,
+      ],
+      providers: [{ provide: CheckoutService, useClass: MockCheckoutService }],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(OrderConfirmationItemsComponent);
+    component = fixture.componentInstance;
+  });
+
+  it('should create', () => {
+    component.ngOnInit();
+    expect(component).toBeTruthy();
+  });
+
+  it('should display items', () => {
+    const items = () => fixture.debugElement.query(By.css('cx-cart-item-list'));
+    component.ngOnInit();
+    fixture.detectChanges();
+    expect(items()).toBeTruthy();
+  });
+});
