+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-items/order-confirmation-items.component.ts
@@ -1,0 +1,27 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
+import { CheckoutService, Order } from '@spartacus/core';
+import { Observable } from 'rxjs';
+
+@Component({
+  selector: 'cx-order-confirmation-items',
+  templateUrl: './order-confirmation-items.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class OrderConfirmationItemsComponent implements OnInit, OnDestroy {
+  order$: Observable<Order>;
+
+  constructor(protected checkoutService: CheckoutService) {}
+
+  ngOnInit() {
+    this.order$ = this.checkoutService.getOrderDetails();
+  }
+
+  ngOnDestroy() {
+    this.checkoutService.clearCheckoutData();
+  }
+}
