+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-overview/order-confirmation-overview.component.html
@@ -1,0 +1,40 @@
+<div class="cx-order-review-summary" *ngIf="(order$ | async) as order">
+  <div class="container">
+    <div class="row">
+      <div class="col-sm-12 col-md-4 col-lg-3">
+        <div class="summary-card">
+          <cx-card
+            [content]="getAddressCardContent(order.deliveryAddress) | async"
+          ></cx-card>
+        </div>
+      </div>
+
+      <div class="col-sm-12 col-md-4 col-lg-3">
+        <div class="summary-card">
+          <cx-card
+            [content]="
+              getBillingAddressCardContent(order.paymentInfo.billingAddress)
+                | async
+            "
+          ></cx-card>
+        </div>
+      </div>
+
+      <div class="col-sm-12 col-md-4 col-lg-3">
+        <div class="summary-card">
+          <cx-card
+            [content]="getDeliveryModeCardContent(order.deliveryMode) | async"
+          ></cx-card>
+        </div>
+      </div>
+
+      <div class="col-sm-12 col-md-4 col-lg-3">
+        <div class="summary-card">
+          <cx-card
+            [content]="getPaymentInfoCardContent(order.paymentInfo) | async"
+          ></cx-card>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
