+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-thank-you-message/order-confirmation-thank-you-message.component.ts
@@ -1,0 +1,28 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
+import { CheckoutService, Order } from '@spartacus/core';
+import { Observable } from 'rxjs';
+
+@Component({
+  selector: 'cx-order-confirmation-thank-you-message',
+  templateUrl: './order-confirmation-thank-you-message.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class OrderConfirmationThankYouMessageComponent
+  implements OnInit, OnDestroy {
+  order$: Observable<Order>;
+
+  constructor(protected checkoutService: CheckoutService) {}
+
+  ngOnInit() {
+    this.order$ = this.checkoutService.getOrderDetails();
+  }
+
+  ngOnDestroy() {
+    this.checkoutService.clearCheckoutData();
+  }
+}
