+++ projects/storefrontlib/src/cms-components/order-confirmation/components/order-confirmation-totals/order-confirmation-totals.component.ts
@@ -1,0 +1,27 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
+import { CheckoutService, Order } from '@spartacus/core';
+import { Observable } from 'rxjs';
+
+@Component({
+  selector: 'cx-order-confirmation-totals',
+  templateUrl: './order-confirmation-totals.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class OrderConfirmationTotalsComponent implements OnInit, OnDestroy {
+  order$: Observable<Order>;
+
+  constructor(protected checkoutService: CheckoutService) {}
+
+  ngOnInit() {
+    this.order$ = this.checkoutService.getOrderDetails();
+  }
+
+  ngOnDestroy() {
+    this.checkoutService.clearCheckoutData();
+  }
+}
