+++ projects/storefrontlib/src/cms-components/order-confirmation/guards/index.ts
@@ -1,0 +1,1 @@
+export * from './order-confirmation.guard';
