+++ projects/storefrontlib/src/cms-components/order-confirmation/guards/order-confirmation.guard.ts
@@ -1,28 +1,28 @@
 import { Injectable } from '@angular/core';
-import { CanActivate } from '@angular/router';
+import { CanActivate, UrlTree, Router } from '@angular/router';
 
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { RoutingService, CheckoutService } from '@spartacus/core';
+import { CheckoutService, SemanticPathService } from '@spartacus/core';
 
 @Injectable({
   providedIn: 'root',
 })
-export class OrderConfirmationPageGuard implements CanActivate {
+export class OrderConfirmationGuard implements CanActivate {
   constructor(
     private checkoutService: CheckoutService,
-    private routingService: RoutingService
+    private router: Router,
+    private semanticPathService: SemanticPathService
   ) {}
 
-  canActivate(): Observable<boolean> {
+  canActivate(): Observable<boolean | UrlTree> {
     return this.checkoutService.getOrderDetails().pipe(
       map(orderDetails => {
         if (orderDetails && Object.keys(orderDetails).length !== 0) {
           return true;
         } else {
-          this.routingService.go({ cxRoute: 'orders' });
-          return false;
+          return this.router.parseUrl(this.semanticPathService.get('orders'));
         }
       })
     );
