+++ projects/storefrontlib/src/cms-components/order-confirmation/index.ts
@@ -1,0 +1,3 @@
+export * from './order-confirmation.module';
+export * from './components/index';
+export * from './guards/index';
