--- projects/storefrontlib/src/cms-pages/order-confirmation-page/order-confirmation-page.component.html
@@ -1,9 +1,0 @@
-<cx-page-layout>
-  <!-- 
-    TODO: as long as order confirmation isn't a cms component we render
-    the hard-coded version to  OrderConfirmationOverviewComponent
-  -->
-  <ng-template cxOutletRef="OrderConfirmationOverviewComponent">
-    <cx-order-confirmation></cx-order-confirmation>
-  </ng-template>
-</cx-page-layout>
