--- projects/storefrontlib/src/cms-pages/order-confirmation-page/order-confirmation-page.component.ts
@@ -1,8 +1,0 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
-
-@Component({
-  selector: 'cx-order-confirmation-page',
-  templateUrl: './order-confirmation-page.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class OrderConfirmationPageComponent {}
