+++ projects/storefrontlib/src/cms-structure/routing/default-routing-config.ts
@@ -32,6 +32,9 @@ export const defaultStorefrontRoutesConfig: RoutesConfig = {
     paths: ['my-account/order/:orderCode'],
     paramsMapping: { orderCode: 'code' },
   },
+  orders: {
+    paths: ['my-account/orders'],
+  },
 };
 
 export const defaultRoutingConfig: RoutingConfig = {
