+++ projects/storefrontlib/src/storefront.module.ts
@@ -12,7 +12,6 @@ import {
 } from '@spartacus/core';
 import { CmsLibModule } from './cms-components/index';
 import { CartPageModule } from './cms-pages/cart-page/cart-page.module';
-import { OrderConfirmationPageModule } from './cms-pages/order-confirmation-page/order-confirmation-page.module';
 import { ProductDetailsPageModule } from './cms-pages/product-details-page/product-details-page.module';
 import { ProductListingPageModule } from './cms-pages/product-listing-page/product-listing-page.module';
 import { CmsModule } from './cms-structure/cms.module';
@@ -41,7 +40,6 @@ import { LayoutModule } from './layout/layout.module';
     ProductDetailsPageModule,
     ProductListingPageModule,
     CartPageModule, // as longs as we do not have #2661 in place we need a specific cart page module
-    OrderConfirmationPageModule, // temp add here till it gets removed by #2691
   ],
   providers: [...provideConfigFromMetaTags()],
   declarations: [],
