+++ projects/storefrontstyles/scss/_sparta-theme.scss
@@ -18,7 +18,9 @@ $selectors: header, cx-storefront, cx-site-context-selector, cx-breadcrumb,
   cx-checkout-progress-mobile-top, cx-checkout-progress-mobile-bottom,
   cx-delivery-mode, cx-shipping-address, cx-billing-address-form,
   cx-payment-form, cx-review-submit, cx-address-form,
-  cx-suggested-addresses-dialog, cx-order-confirmation, cx-promotions,
+  cx-suggested-addresses-dialog, cx-order-confirmation-items,
+  cx-order-confirmation-totals, cx-order-confirmation-overview,
+  cx-order-confirmation-thank-you-message, cx-promotions,
   cx-order-detail-headline, cx-order-detail-items, cx-order-detail-shipping,
   cx-order-detail-totals, cx-product-view, cx-payment-method, cx-payment-methods,
   cx-close-account, cx-store-finder-list-item, cx-store-finder-grid,
