+++ projects/storefrontstyles/scss/components/checkout/_index.scss
@@ -1,4 +1,4 @@
 @import './cart/index';
 @import './steps/index';
-@import './order-confirmation';
+@import './order-confirmation/index';
 @import './promotions';
