--- projects/storefrontstyles/scss/components/checkout/_order-confirmation.scss
@@ -1,105 +1,0 @@
-@import '../../theme';
-
-%cx-order-confirmation {
-  .cx-order-confirmation-message {
-    text-align: center;
-    padding: 2.5rem;
-
-    h2 {
-      font-weight: $font-weight-normal;
-    }
-
-    .btn-link {
-      font-size: $small-font-size;
-      font-weight: $font-weight-bold;
-      text-transform: uppercase;
-    }
-  }
-
-  .cx-page-title {
-    color: var(--cx-g-color-text);
-  }
-
-  .cx-order-review-summary {
-    background-color: var(--cx-g-color-background);
-    border-width: 0 0 1px 0;
-    border-style: solid;
-    border-color: var(--cx-g-color-light);
-
-    @include media-breakpoint-down(md) {
-      background-color: var(--cx-g-color-inverse);
-    }
-
-    .summary-card {
-      @include media-breakpoint-down(sm) {
-        background-color: var(--cx-g-color-inverse);
-        border-width: 1px;
-        border-style: solid;
-        border-color: var(--cx-g-color-light);
-        margin: 0.625rem 0;
-      }
-    }
-
-    .container {
-      padding: 0;
-      @include media-breakpoint-down(md) {
-        max-width: 100%;
-        min-width: 100%;
-        padding: 0 1.25rem;
-      }
-      @include media-breakpoint-down(sm) {
-        padding: 1.25rem;
-      }
-    }
-  }
-
-  .cx-order-items {
-    &.container {
-      @include media-breakpoint-down(md) {
-        max-width: 100%;
-        min-width: 100%;
-        padding: 0;
-      }
-      @include media-breakpoint-down(sm) {
-        max-width: 100%;
-        min-width: 100%;
-        padding: 0;
-      }
-    }
-  }
-
-  .cx-order-items-header {
-    @include type('4');
-    padding: 1.375rem 0;
-    margin: 0;
-    border-width: 0 0 1px 0;
-    border-style: solid;
-    border-color: var(--cx-g-color-light);
-    @include media-breakpoint-down(md) {
-      max-width: 100%;
-      min-width: 100%;
-      padding-left: 2.5rem;
-    }
-    @include media-breakpoint-down(sm) {
-      max-width: 100%;
-      min-width: 100%;
-      padding-left: 1rem;
-    }
-  }
-
-  .cx-order-summary {
-    padding-right: 0;
-
-    &.container {
-      @include media-breakpoint-down(md) {
-        max-width: 100%;
-        min-width: 100%;
-        padding-right: 1.625rem;
-      }
-
-      @include media-breakpoint-down(sm) {
-        padding: 0;
-      }
-    }
-  }
-}
