+++ projects/storefrontstyles/scss/components/checkout/order-confirmation/_index.scss
@@ -1,0 +1,4 @@
+@import './order-confirmation-items';
+@import './order-confirmation-overview';
+@import './order-confirmation-thank-you-message';
+@import './order-confirmation-totals';
