+++ projects/storefrontstyles/scss/components/checkout/order-confirmation/_order-confirmation-overview.scss
@@ -1,0 +1,36 @@
+@import '../../../theme';
+
+%cx-order-confirmation-overview {
+  .cx-order-review-summary {
+    background-color: var(--cx-background-color, var(--cx-g-color-background));
+    border-width: var(--cx-border-width, 0 0 1px 0);
+    border-style: var(--cx-border-style, solid);
+    border-color: var(--cx-border-color, var(--cx-g-color-light));
+
+    @include media-breakpoint-down(md) {
+      background-color: var(--cx-background-color, var(--cx-g-color-inverse));
+    }
+
+    .summary-card {
+      @include media-breakpoint-down(sm) {
+        background-color: var(--cx-background-color, var(--cx-g-color-inverse));
+        border-width: var(--cx-border-width, 1px);
+        border-style: var(--cx-border-style, solid);
+        border-color: var(--cx-border-color, var(--cx-g-color-light));
+        margin: var(--cx-margin, 0.625rem 0);
+      }
+    }
+
+    .container {
+      padding: var(--cx-padding, 0);
+      @include media-breakpoint-down(md) {
+        max-width: var(--cx-max-width, 100%);
+        min-width: var(--cx-min-width, 100%);
+        padding: var(--cx-padding, 0 1.25rem);
+      }
+      @include media-breakpoint-down(sm) {
+        padding: var(--cx-padding, 1.25rem);
+      }
+    }
+  }
+}
