+++ projects/storefrontstyles/scss/components/checkout/order-confirmation/_order-confirmation-totals.scss
@@ -1,0 +1,19 @@
+@import '../../../theme';
+
+%cx-order-confirmation-totals {
+  .cx-order-summary {
+    padding-right: var(--cx-padding, 0);
+
+    &.container {
+      @include media-breakpoint-down(md) {
+        max-width: var(--cx-max-width, 100%);
+        min-width: var(--cx-min-width, 100%);
+        padding-right: var(--cx-padding, 1.625rem);
+      }
+
+      @include media-breakpoint-down(sm) {
+        padding: var(--cx-padding, 0);
+      }
+    }
+  }
+}
